fix: do not pass react nodes
If we pass react nodes, we will have
issues with react's hook rules when
trying to `clearContextMenu()` to hide the
context menu upon the user clicking
an option.

This commit fixes that and puts the user
in a situation where they
don't have to worry about clearing
the context menu state.